hw 9 + hw10

+++ app/src/main/java/vlasov/ru/androidfundamentalsproject/data/locale/LocalDataSource.kt
@@ -5,5 +5,5 @@ import vlasov.ru.androidfundamentalsproject.models.Movie
 interface LocalDataSource {
     suspend fun loadMovies() : List<Movie>
     suspend fun insertMovies(movies : List<Movie>)
-    suspend fun loadMovie(id : Int): Movie
+    suspend fun loadMovie(id : Long): Movie
 }
